+++ go.mod
@@ -1,3 +1,16 @@
-module did-this
+module github.com/nshttpd/did-this
 
 go 1.17
+
+require (
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/cobra v1.3.0
+	go.etcd.io/bbolt v1.3.6
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/sys v0.0.0-20211205182925-97ca703d548d // indirect
+)
