+++ internal/commands/add.go
@@ -27,75 +27,44 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-package cmd
+package commands
 
 import (
-	"fmt"
-	"time"
+	"encoding/binary"
+	bolt "go.etcd.io/bbolt"
 
-	"os"
+	"strings"
 
-	"github.com/coreos/bbolt"
 	"github.com/spf13/cobra"
 )
 
-const RETENTION = 30
+// addCmd represents the add command
+var addCmd = &cobra.Command{
+	Use:   "add",
+	Short: "Add a done task for today",
+	Long: `Add in a done task for the day that you want to be able to report on tomorrow
+to say that you've actually done something.
 
-var purgeRetentionVar int
+	did-this add "I did something sort of cool today"
 
-// purgeCmd represents the purge command
-var purgeCmd = &cobra.Command{
-	Use:   "purge",
-	Short: "purge out old completed stored tasks",
-	Long: `Loop through the database and purge out old daily buckets
-of data. Default retention period is 30 days. Can be overridden on the
-command line.
-
-	did-this purge
-
-this is destructive and no backup of data is made.`,
+`,
 	Run: func(cmd *cobra.Command, args []string) {
+		cfg.Db.Update(func(tx *bolt.Tx) error {
+			b := tx.Bucket(cfg.CurrentDate())
+			id, _ := b.NextSequence()
 
-		var purge [][]byte
-		now := time.Now()
-
-		err := cfg.Db.View(func(tx *bolt.Tx) error {
-			return tx.ForEach(func(name []byte, _ *bolt.Bucket) error {
-				d, err := time.Parse("2006-01-02", string(name))
-				if err != nil {
-					return fmt.Errorf("error parsing time from bucket : %s", string(name))
-				}
-
-				if d.Before(now.AddDate(0, 0, -purgeRetentionVar)) {
-					purge = append(purge, name)
-				}
-				return nil
-			})
+			return b.Put(itob(id), []byte(strings.Join(args, " ")))
 		})
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
 
-		if len(purge) > 0 {
-			err := cfg.Db.Update(func(tx *bolt.Tx) error {
-				for _, b := range purge {
-					if e := tx.DeleteBucket(b); e != nil {
-						return e
-					}
-					fmt.Printf("purged : %s\n", string(b))
-				}
-				return nil
-			})
-			if err != nil {
-				fmt.Println(err)
-			}
-		}
 	},
 }
 
+func itob(v uint64) []byte {
+	b := make([]byte, 8)
+	binary.BigEndian.PutUint64(b, v)
+	return b
+}
+
 func init() {
-	rootCmd.AddCommand(purgeCmd)
-	purgeCmd.Flags().IntVar(&purgeRetentionVar, "retention", RETENTION,
-		"retention days for completed tasks")
+	rootCmd.AddCommand(addCmd)
 }
